api/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
PyJWT==2.4.0
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.2
